requirements.txt
llama-cpp-python
langchain
langchain-community
sentence-transformers
chromadb
pypdf
#requestspydantic
tqdm
jupyter

langchain-huggingface
faiss-cpu
pytorch
torch
torchvision
torchaudio
matplotlib
fastapi
